requirements.txt
python-dotenv==1.0.0
Flask==3.0.0
pydantic===2.4.2
Flask-Cors==4.0.0
pytest==7.3.1
pytest-cov==4.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
email-validator==2.1.0.post1
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2